+++ magik-checks/src/main/java/nl/ramsolutions/sw/magik/checks/CheckList.java
@@ -7,6 +7,7 @@ import nl.ramsolutions.sw.magik.checks.checks.DuplicateMethodInFileCheck;
 import nl.ramsolutions.sw.magik.checks.checks.EmptyBlockCheck;
 import nl.ramsolutions.sw.magik.checks.checks.ExemplarSlotCountCheck;
 import nl.ramsolutions.sw.magik.checks.checks.FileMethodCountCheck;
+import nl.ramsolutions.sw.magik.checks.checks.FileMustStartWithPackageStatementCheck;
 import nl.ramsolutions.sw.magik.checks.checks.FileNotInLoadListCheck;
 import nl.ramsolutions.sw.magik.checks.checks.ForbiddenCallCheck;
 import nl.ramsolutions.sw.magik.checks.checks.ForbiddenGlobalUsageCheck;
@@ -68,6 +69,7 @@ public final class CheckList {
         ExemplarSlotCountCheck.class,
         FileMethodCountCheck.class,
         FileNotInLoadListCheck.class,
+        FileMustStartWithPackageStatementCheck.class,
         ForbiddenCallCheck.class,
         ForbiddenGlobalUsageCheck.class,
         ForbiddenInheritanceCheck.class,