+++ CHANGES.md
@@ -27,6 +27,7 @@
 - Add UndefinedVariable check to Sonar way profile.
 - Add ability to override aliases path and environment path in tasks in VSCode extension.
 - Set default value for `whitelist` in VariableNamingCheck.
+- Add FileMustStartWithPackageStatement check to test if a file starts with a `_package`-statement.
 - Several fixes.
 
 ### Breaking changes (reiterated from above)
@@ -38,6 +39,7 @@
 - Add SimplifyIf check to Sonar way profile. This might result in more issues.
 - Add UnsafeEvaluateInvocation check to Sonar way profile. This might result in more issues.
 - Add UndefinedVariable check to Sonar way profile. This might result in more issues.
+- Add FileMustStartWithPackageStatement check to test if a file starts with a `_package`-statement. This might result in more issues.
 
 ## 0.10.1 (2024-08-14)